srcs/Authenticator/Auth_service/requirements.txt
Django==4.2
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
django-cors-headers
django-environ
python-dotenv
requests